refactored: moved inidef/porttype.go to device/porttype.go.
InputPortType and its values are used in device package only.
