+++ device/porttype.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package inidef
+package device
 
 // InputPortType defines port type used by gateway
 type InputPortType int
